requirements.txt
asgiref==3.7.2
Babel==2.12.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.33.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
django-appconf==1.0.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.1.0
django-crispy-forms==1.14.0
django-money==3.1.0
django-recaptcha==3.0.0
django-select2==8.1.2
djangorestframework==3.14.0
docxcompose==1.4.0
docxtpl==0.16.7
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
oauthlib==3.2.2
Pillow==9.5.0
py-moneyed==2.0
pycparser==2.21
PyJWT==2.7.0
python-docx==0.8.11
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stripe==5.0.0
typing_extensions==4.6.3
tzdata==2022.6
urllib3==1.26.16
Werkzeug==2.2.2
whitenoise==6.5.0
